+++ AdministratorPHPFiles/editQuizQuestion.php
@@ -117,17 +117,26 @@ if (isset($_POST)) {
             $images[] = $iRow['name'];
         }
 
-        if ($images[0] != "None") {
-            for ($i = 0; $i < count($images); $i++) {
-                $quizImgLocation = "../css/img/quizImg/quiz" . strval($quizID) . "/" . $images[$i];
-                unlink($quizImgLocation);
-            }//end of images delete loop
-        }
 
         $deleteImagesQuery = "DELETE FROM question_image WHERE question_id = $questionID";
         $deleteImagesResult = mysqli_query($link, $deleteImagesQuery) or die(mysqli_error($link));
 
         if ($deleteImagesQuery) {
+            $existingImgQuery = "SELECT name FROM question_image";
+
+            $existingImgResult = mysqli_query($link, $existingImgQuery) or die(mysqli_error($link));
+
+            while ($eIRow = mysqli_fetch_assoc($existingImgResult)) {
+                $existingImgArr[] = $eIRow['name'];
+            }
+
+            $imagesToDelete = array_values(array_diff($images, $existingImgArr));
+            if ($images[0] != "None" && count($imagesToDelete) != 0) {
+                for ($i = 0; $i < count($imagesToDelete); $i++) {
+                    $quizImgLocation = "../css/img/quizImg/quiz" . strval($quizID) . "/" . $imagesToDelete[$i];
+                    unlink($quizImgLocation);
+                }//end of images delete loop
+            }
             if (isset($_FILES['files'])) {
 
                 $countfiles = count($_FILES['files']['name']);
